# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loadout HTTP API",
    "version": "0.2.0",
    "description": "实际Go app/identity/运维路由。网页会话与ldt_网关令牌分离；业务写请求校验Origin；外部支付目前不可用。Body总大小app默认16KiB、技能发布50MiB（另验32文件/单文件256KiB）、identity4KiB；字符串长度限制以Go UTF-8字节校验为准，JSON Schema maxLength仅作客户端提示。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源服务根地址"
    }
  ],
...
  "security": [
    {
      "CookieSession": []
    }
  ],
  "paths": {
...
    "/mcp": {
      "post": {
        "operationId": "post_mcp",
        "summary": "MCP Streamable HTTP / JSON-RPC",
        "tags": ["MCP"],
        "description": "官方MCP Go SDK stateless + JSONResponse。调用initialize、notifications/initialized、tools/list、tools/call等标准MCP方法；不由本OpenAPI重新定义全部MCP协议。仅POST，认证ldt_；请求体上限1MiB。",
        "responses": {
          "200": {
            "description": "SDK JSON-RPC响应；工具失败可为result.isError=true。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "202": {
            "description": "SDK已接受通知，无响应体"
          },
          "400": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "405": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "413": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "503": {
            "description": "HTTP传输、鉴权或SDK错误；不保证为REST Error结构。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "GatewayToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "const": "2.0"
                  },
                  "id": {
                    "type": ["string", "integer", "null"]
                  },
                  "method": {
                    "type": "string"
                  },
                  "params": {}
                },
                "required": ["jsonrpc", "method"]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "MCP-Protocol-Version",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "后续请求按initialize协商版本携带。"
          },
          {
            "name": "Accept",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "application/json, text/event-stream"
          }
        ]
      },
      "get": {
        "operationId": "get_mcp_unsupported",
        "summary": "此部署不提供MCP流式GET/DELETE会话操作",
        "tags": ["MCP"],
        "security": [
          {
            "GatewayToken": []
          }
        ],
        "responses": {
          "401": {
            "description": "缺少或无效Bearer",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "description": "已认证请求仅允许POST",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                },
                "example": "POST"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_mcp_unsupported",
        "summary": "此部署不提供MCP流式GET/DELETE会话操作",
        "tags": ["MCP"],
        "security": [
          {
            "GatewayToken": []
          }
        ],
        "responses": {
          "401": {
            "description": "缺少或无效Bearer",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "description": "已认证请求仅允许POST",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                },
                "example": "POST"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_session",
        "description": "HttpOnly、SameSite=Lax，默认900秒opaque AT；生产Secure。旧7天session在原期限内兼容。"
      },
      "GatewayToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ldt_ token",
        "description": "网页会话不能替代此令牌。"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loadout_refresh"
      }
    },
...
  }
}